Extract Organization into it's own file
