+++ lib/everypolitician/popolo.rb
@@ -1,6 +1,7 @@
 require 'everypolitician/popolo/version'
 require 'everypolitician/popolo/collection'
 require 'everypolitician/popolo/people'
+require 'everypolitician/popolo/organization'
 require 'json'
 
 module Everypolitician
@@ -42,28 +43,6 @@ module Everypolitician
       def memberships
         Memberships.new(popolo[:memberships])
       end
-    end
-
-    class Organizations < Collection
-      def initialize(documents)
-        @documents = documents.map { |p| Organization.new(p) }
-      end
-    end
-
-    class Organization
-      attr_reader :document
-
-      def initialize(document)
-        @document = document
-        document.each do |key, value|
-          define_singleton_method(key) { value }
-        end
-      end
-
-      def ==(other)
-        id == other.id
-      end
-      alias eql? ==
     end
 
     class Areas < Collection
